chore(ci): use code.forgejo.org/forgejo/migration-test for migration tests (#8157)
Using https://github.com/go-gitea/test_repo is not necessary as the current API integration tests are not GitHub specific. It makes them less susceptible to transient failures because of throttling.

---

The https://code.forgejo.org/forgejo/migration-test repository was created just now. There still is a chance (although very slim) that someone running the CI will be throttled by code.forgejo.org and the error message suggesting that may be a cause for failure is kept to clear the confusion in case it happens.

Reviewed-on: https://codeberg.org/forgejo/forgejo/pulls/8157
Reviewed-by: Gusted <gusted@noreply.codeberg.org>
Co-authored-by: Earl Warren <contact@earl-warren.org>
Co-committed-by: Earl Warren <contact@earl-warren.org>